requirements.txt
autopep8==2.0.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
google-api-core==2.10.2
google-auth==2.14.0
google-cloud-speech==2.16.2
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
idna==3.4
proto-plus==1.22.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
requests==2.28.1
rsa==4.9
six==1.16.0
tomli==2.0.1
urllib3==1.26.12
